+++ src/main/kotlin/com/quickref/plugin/action/DiffSourceAction.kt
@@ -11,7 +11,6 @@ import com.quickref.plugin.extension.downloadFile
 import com.quickref.plugin.extension.guessFileName
 import com.quickref.plugin.extension.isAndroidFrameworkClass
 import com.quickref.plugin.extension.isSupport
-import com.quickref.plugin.extension.packageName
 import com.quickref.plugin.extension.pathname
 import com.quickref.plugin.version.AndroidVersion
 import com.quickref.plugin.widget.AndroidVersionsPopView