+++ src/main/kotlin/com/quickref/plugin/download/engine/AOSPDownload.kt
@@ -10,7 +10,7 @@ class AOSPDownload : FileDownload(Source.AOSPMirror) {
         // android-14.0.0_r69/identity/java/android/security/identity/AccessControlProfile.java
         // ?format=TEXT"| base64 --decode > AccessControlProfile.java
         private const val DOWNLOAD_PATH =
-            "https://android.googlesource.com/platform/frameworks/base.git/+/refs/tags/%s%s?format=TEXT"
+            "https://android.googlesource.com/platform/frameworks/base.git/+/refs/tags/%s/%s?format=TEXT"
     }
 
     override fun baseDownloadURL(): String {