+++ src/test/java/com/quickref/plugin/db/DBGenerator.kt
@@ -2,19 +2,42 @@
 
 package com.quickref.plugin.db
 
-import com.quickref.plugin.git.RepoType
-import com.quickref.plugin.git.getAOSPRootPath
 import com.quickref.plugin.version.AndroidVersion
 import com.quickref.plugin.version.Source
 import org.junit.Before
 import org.junit.Test
 import java.io.File
+import java.sql.DriverManager
+import java.sql.SQLException
+
 
 class DBGenerator {
 
+    @Test
+    fun aospDirTest() {
+        AndroidVersion
+            .getVersionSource(Source.AOSPMirror)
+            .versionPairs()
+            .entries
+            .forEach { entry ->
+                val branch = entry.value
+
+                val javaSource = File("${RepoType.AOSP_BACKUP.getAOSPRootPath()}/$branch/java")
+                val cppSource = File("${RepoType.AOSP_BACKUP.getAOSPRootPath()}/$branch/cpp")
+
+                val versionNumber = AndroidVersion.getBuildNumber(branch)
+                println("branch:$branch version:$versionNumber")
+                println("javaSource:$javaSource")
+                println("cppSource:$cppSource")
+
+                assert(versionNumber != 0)
+                assert(javaSource.exists())
+                assert(cppSource.exists())
+            }
+    }
+
     private val javaFileSQ = File("src/main/sqldelight/com/quickref/plugin/db/JavaFileMapping.sq")
     private val nativeFileSQ = File("src/main/sqldelight/com/quickref/plugin/db/NativeFileMapping.sq")
-    private val nativeMethodSQ = File("src/main/sqldelight/com/quickref/plugin/db/NativeMethodMapping.sq")
 
     @Before
     fun setup() {
@@ -42,111 +65,51 @@ class DBGenerator {
         val nativeFileTable = """
         CREATE TABLE NativeFileMapping (
           psiFile TEXT NOT NULL,
+          jniClass TEXT NOT NULL,
           version INTEGER DEFAULT 0,
           path TEXT NOT NULL
         );
 
-        CREATE INDEX NativeFileMapping_path ON NativeFileMapping(psiFile,version);
+        CREATE INDEX NativeFileMapping_jniClass ON NativeFileMapping(psiFile,jniClass,version);
 
         selectAll:
-        SELECT psiFile,version,path FROM NativeFileMapping;
+        SELECT psiFile,jniClass,version,path FROM NativeFileMapping;
 
         count:
         SELECT count(*) FROM NativeFileMapping;
 
         getNativeFile:
-        SELECT path FROM NativeFileMapping WHERE psiFile=:file AND version >=:version ORDER BY version ASC LIMIT 1;
-    """.trimIndent()
-
-        val nativeMethodTable = """
-        CREATE TABLE NativeMethodMapping (
-          file TEXT NOT NULL,
-          jniMethod TEXT NOT NULL,
-          nativeMethod TEXT NOT NULL,
-          version INTEGER DEFAULT 0,
-          jniLine INTEGER DEFAULT 0,
-          defLine INTEGER DEFAULT 0
-        );
-
-        CREATE INDEX NativeMapping_line ON NativeMethodMapping(file,jniMethod,nativeMethod,version,jniLine,defLine);
-
-        getNativeMethod:
-        SELECT * FROM NativeMethodMapping WHERE file=:key AND jniMethod=:method AND version >=:version ORDER BY version ASC LIMIT 1;
-
-        count:
-        SELECT count(*) FROM NativeMethodMapping;
-
-        getNativeMethodByName:
-        SELECT * FROM NativeMethodMapping WHERE jniMethod=:name LIMIT 1;
+        SELECT path FROM NativeFileMapping WHERE jniClass=:clazz AND version >=:version ORDER BY version ASC LIMIT 1;
     """.trimIndent()
 
         javaFileSQ.writeText(javaFileTable)
         nativeFileSQ.writeText(nativeFileTable)
-        nativeMethodSQ.writeText(nativeMethodTable)
+    }
+
+    enum class RepoType {
+        AOSP_BACKUP,
+        EXTERNAL_SKIA,
+    }
+
+    private fun RepoType.getAOSPRootPath(): String {
+        return when (this) {
+            RepoType.AOSP_BACKUP -> System.getenv("AOSP_BACKUP")
+            RepoType.EXTERNAL_SKIA -> System.getenv("EXTERNAL_SKIA")
+        }
+    }
+
+    private fun excludeDir(file: File): Boolean {
+        val path = file.absolutePath
+        return file.isDirectory ||
+            path.contains("/tools/") ||
+            path.contains("/apct-tests/") ||
+            path.contains("/test-runner/") ||
+            path.contains("/include/")
     }
 
     @Test
-    fun aospDirTest() {
-        AndroidVersion
-            .getVersionSource(Source.AOSPMirror)
-            .versionPairs()
-            .entries
-            .forEach { entry ->
-                val branch = entry.value
-                val javaSource = File("${getAOSPRootPath(RepoType.AOSP_BASE)}/aosp_base/java_files/$branch")
-                val cppSource = File("${getAOSPRootPath(RepoType.AOSP_BASE)}/aosp_base/cpp_files/$branch")
-
-                val versionNumber = AndroidVersion.getBuildNumber(branch)
-                println("branch:$branch version:$versionNumber")
-
-                assert(versionNumber != 0)
-                assert(javaSource.exists())
-                assert(cppSource.exists())
-            }
-    }
-
-    @Test
-    fun generatorDisSQL() {
-        val javaSQL = """
-        DELETE FROM JavaFileMapping AS a
-        WHERE(
-            a.psiFile,
-            a.version,
-            a.path
-        ) IN (SELECT psiFile, version, path FROM JavaFileMapping GROUP BY psiFile, version, path HAVING count(*) > 1)
-        AND rowid NOT IN (SELECT min (rowid) FROM JavaFileMapping GROUP BY psiFile, version, path HAVING count(*)>1);
-        """.trimIndent()
-
-        val cppSQL = """
-        DELETE FROM NativeFileMapping AS a
-            WHERE(
-                a.psiFile,
-                a.version,
-                a.path
-            ) IN (SELECT psiFile, version, path FROM NativeFileMapping GROUP BY psiFile, version, path HAVING count(*) > 1)
-        AND rowid NOT IN (SELECT min (rowid) FROM NativeFileMapping GROUP BY psiFile, version, path HAVING count(*)>1);
-        DELETE FROM NativeMethodMapping AS a
-        WHERE(
-            a.file,
-            a.jniMethod,
-            a.nativeMethod,
-            a.version,
-            a.jniLine,
-            a.defLine
-        ) IN (SELECT file, jniMethod, nativeMethod, version, jniLine, defLine FROM NativeMethodMapping GROUP BY file, jniMethod, nativeMethod, version, jniLine, defLine HAVING count(*) > 1)
-        AND rowid NOT IN (SELECT min (rowid) FROM NativeMethodMapping GROUP BY file, jniMethod, nativeMethod, version, jniLine, defLine HAVING count(*)>1);
-        """.trimIndent()
-
-        val javaDisSQL = File("src/test/resources/JavaDis.sql")
-        val cppDisSQL = File("src/test/resources/NativeDis.sql")
-        javaDisSQL.writeText(javaSQL)
-        cppDisSQL.writeText(cppSQL)
-    }
-
-    @Test
-    fun generator() {
-
-        val javaSourceSQL = File("src/test/resources/JavaSource.sql")
+    fun generatorJava() {
+        val javaSourceSQL = File("src/test/resources/JavaFileMapping.txt")
         javaSourceSQL.writeText("")
         AndroidVersion
             .getVersionSource(Source.AOSPMirror)
@@ -154,11 +117,12 @@ class DBGenerator {
             .entries
             .forEach { entry ->
                 val branch = entry.value
-                val javaSource = File("${getAOSPRootPath(RepoType.AOSP_BASE)}/aosp_base/java_files/$branch")
+                val javaSource = File("${RepoType.AOSP_BACKUP.getAOSPRootPath()}/$branch/java")
                 val versionNumber = AndroidVersion.getBuildNumber(branch)
-                javaSource.walk()
+                javaSource
+                    .walk()
                     .maxDepth(depth = 16)
-                    .filter { file -> file.isDirectory.not() }
+                    .filter { file -> excludeDir(file).not() }
                     .forEach { file ->
                         // java key
                         val javaFileBuilder = StringBuilder()
@@ -168,12 +132,12 @@ class DBGenerator {
                         }
                     }
             }
+    }
 
-
-        val cppSourceSQL = File("src/test/resources/CppSource.sql")
-        val cppMethodSQL = File("src/test/resources/CppMethod.sql")
+    @Test
+    fun generatorCpp() {
+        val cppSourceSQL = File("src/test/resources/NativeFileMapping.txt")
         cppSourceSQL.writeText("")
-        cppMethodSQL.writeText("")
 
         AndroidVersion
             .getVersionSource(Source.AOSPMirror)
@@ -181,26 +145,175 @@ class DBGenerator {
             .entries
             .forEach { entry ->
                 val branch = entry.value
-                val cppSource = File("${getAOSPRootPath(RepoType.AOSP_BASE)}/aosp_base/cpp_files/$branch")
+                val cppSource = File("${RepoType.AOSP_BACKUP.getAOSPRootPath()}/$branch/cpp")
                 val versionNumber = AndroidVersion.getBuildNumber(branch)
                 cppSource.walk()
-                    .maxDepth(depth = 10)
-                    .filter { file -> file.isDirectory.not() }
+                    .maxDepth(depth = 16)
+                    .filter { file -> excludeDir(file).not() }
                     .filter { file -> file.readLines().any { line -> line.contains("JNINativeMethod") } }
                     .forEach { file ->
-                        // native key
-                        val nativeFileBuilder = StringBuilder()
-                        val nativeMethodBuilder = StringBuilder()
-                        readNativeFileLines(file, cppSource, versionNumber, nativeFileBuilder, nativeMethodBuilder)
-                        if (nativeFileBuilder.isNotEmpty()) {
-                            cppSourceSQL.appendText(nativeFileBuilder.toString())
-                        }
-                        if (nativeMethodBuilder.isNotEmpty()) {
-                            cppMethodSQL.appendText(nativeMethodBuilder.toString())
+                        // AndroidRuntime::registerNativeMethods(env, "android/graphics/Canvas", gMethods, NELEM(gMethods));
+                        // RegisterMethodsOrDie(env, "android/graphics/Canvas", gMethods, NELEM(gMethods));
+                        // env->FindClass("android/drm/DrmManagerClient")
+                        val jniClasses = file.readLines()
+                            .mapNotNull { line ->
+                                if (line.count { it == '"' } == 2) {
+                                    try {
+                                        when {
+                                            // static const char* const kClassPathName = "android/media/AudioTrack";
+                                            line.contains("kClassPathName = \"") -> {
+                                                line.substring(
+                                                    startIndex = line.indexOf("\"") + 1,
+                                                    endIndex = line.lastIndexOf("\"")
+                                                )
+                                            }
+
+                                            line.contains("env->FindClass(\"android") -> {
+                                                // try find jni class
+                                                val archer = "env->FindClass(\""
+                                                val className = line.substring(
+                                                    startIndex = line.indexOf(archer) + archer.length,
+                                                    endIndex = line.indexOf("\")")
+                                                )
+                                                val shortName = className.substring(className.lastIndexOf("/") + 1)
+                                                val getCppName = file.nameWithoutExtension
+                                                if (getCppName.endsWith(shortName, ignoreCase = true)) {
+                                                    className
+                                                } else {
+                                                    // println("line:$line\nclassName:$className getCppName:$getCppName shortName:$shortName")
+                                                    null
+                                                }
+                                            }
+
+                                            line.contains("registerNativeMethods(") -> {
+                                                line.substring(
+                                                    startIndex = line.indexOf("\"") + 1,
+                                                    endIndex = line.lastIndexOf("\"")
+                                                )
+                                            }
+
+                                            line.contains("RegisterMethodsOrDie(") -> {
+                                                line.substring(
+                                                    startIndex = line.indexOf("\"") + 1,
+                                                    endIndex = line.lastIndexOf("\"")
+                                                )
+                                            }
+
+                                            else -> null
+                                        }
+                                    } catch (throwable: Throwable) {
+                                        println("crash file:$file line:$line ")
+                                        throw throwable
+                                    }
+                                } else {
+                                    null
+                                }
+                            }
+
+                        // try to record native file
+                        //INSERT INTO NativeFileMapping (psiFile,version,path)
+                        //VALUES ("android/graphics/Bitmap.cpp",29,"core/jni/android/graphics/Bitmap.cpp");
+                        val nativePath = file.parent.replace(cppSource.path, "") + "/" + file.name
+                        jniClasses.forEach { jniClass ->
+                            if (jniClass.isNotEmpty()) {
+                                cppSourceSQL.appendText(
+                                    "INSERT INTO NativeFileMapping (psiFile,jniClass,version,path) VALUES (\"${file.name}\",\"$jniClass\",$versionNumber,\"$nativePath\");\n"
+                                )
+                            } else {
+                                println("jniClass:$jniClass version:$versionNumber path:$nativePath ")
+                            }
                         }
                     }
-
             }
+    }
+
+    @Test
+    fun writeDataSetToDB() {
+        val javaSourceSQL = File("src/test/resources/JavaFileMapping.txt")
+        val cppSourceSQL = File("src/test/resources/NativeFileMapping.txt")
+
+        // sqlite db
+        try {
+            DriverManager.getConnection("jdbc:sqlite:src/main/resources/db/QuickRefDB.db").use { connection ->
+                connection.createStatement().use { statement ->
+                    statement.queryTimeout = 30 // set timeout to 30 sec.
+                    // create a new table
+
+                    statement.executeUpdate("DROP TABLE JavaFileMapping;")
+                    statement.executeUpdate("DROP TABLE NativeFileMapping;")
+
+                    statement.executeUpdate(
+                        """
+                        CREATE TABLE IF NOT EXISTS JavaFileMapping (
+                            psiFile TEXT NOT NULL,
+                            version INTEGER DEFAULT 0,
+                            path TEXT NOT NULL
+                        );
+                        """.trimIndent()
+                    )
+                    statement.executeUpdate(
+                        """
+                        CREATE INDEX IF NOT EXISTS JavaFileMapping_path ON JavaFileMapping(psiFile,version);
+                        """.trimIndent()
+                    )
+                    statement.executeUpdate(
+                        """
+                        CREATE TABLE IF NOT EXISTS NativeFileMapping (
+                          psiFile TEXT NOT NULL,
+                          jniClass TEXT NOT NULL,
+                          version INTEGER DEFAULT 0,
+                          path TEXT NOT NULL
+                        );
+                        """.trimIndent()
+                    )
+                    statement.executeUpdate(
+                        """
+                        CREATE INDEX IF NOT EXISTS NativeFileMapping_jniClass ON NativeFileMapping(psiFile,version,path);
+                        """.trimIndent()
+                    )
+                }
+                connection.createStatement().use { statement ->
+                    javaSourceSQL.readLines().forEach { line ->
+                        statement.executeUpdate(line)
+                    }
+                    cppSourceSQL.readLines().forEach { line ->
+                        statement.executeUpdate(line)
+                    }
+                }
+                connection.createStatement().use { statement ->
+                    statement.executeUpdate(
+                        """
+                        DELETE FROM JavaFileMapping AS a
+                        WHERE(
+                            a.psiFile,
+                            a.version,
+                            a.path
+                        ) IN (SELECT psiFile, version, path FROM JavaFileMapping GROUP BY psiFile, version, path HAVING count(*) > 1)
+                        AND rowid NOT IN (SELECT min (rowid) FROM JavaFileMapping GROUP BY psiFile, version, path HAVING count(*)>1);
+                    """.trimIndent()
+                    )
+                    statement.executeUpdate(
+                        """
+                        DELETE FROM NativeFileMapping AS a
+                            WHERE(
+                                a.psiFile,
+                                a.jniClass,
+                                a.version,
+                                a.path
+                            ) IN (SELECT psiFile, jniClass, version, path FROM NativeFileMapping GROUP BY psiFile, jniClass, version, path HAVING count(*) > 1)
+                        AND rowid NOT IN (SELECT min (rowid) FROM NativeFileMapping GROUP BY psiFile, jniClass, version, path HAVING count(*)>1);
+                    """.trimIndent()
+                    )
+                }
+                connection.createStatement().use { statement ->
+                    statement.executeUpdate("VACUUM;")
+                }
+            }
+        } catch (e: SQLException) {
+            // if the error message is "out of memory",
+            // it probably means no database file is found
+            e.printStackTrace(System.err)
+        }
     }
 
     private fun readJavaFileLines(
@@ -225,122 +338,5 @@ class DBGenerator {
         javaFileBuilder.append(
             "INSERT INTO JavaFileMapping (psiFile,version,path) VALUES (\"$javaKey\",$nativeVersion,\"$nativePath\");\n"
         )
-    }
-
-    private fun readNativeFileLines(
-        file: File,
-        source: File,
-        versionNumber: Int,
-        nativeFileBuilder: StringBuilder,
-        nativeMethodBuilder: StringBuilder
-    ) {
-        val packageIndex = file.name.lastIndexOf("_")
-        val nativeKey = if (packageIndex >= 0) {
-            file.name.substring(packageIndex + 1)
-        } else {
-            file.name
-        }
-        val nativePath = file.parent.replace(source.path, "") + "/" + file.name
-        val nativeVersion: Int = versionNumber
-
-        println("key:$nativeKey version:$nativeVersion path:$nativePath ")
-        // try to record native file
-        //INSERT INTO NativeFileMapping (psiFile,version,path)
-        //VALUES ("android/graphics/Bitmap.cpp",29,"core/jni/android/graphics/Bitmap.cpp");
-        nativeFileBuilder.append(
-            "INSERT INTO NativeFileMapping (psiFile,version,path) VALUES (\"$nativeKey\",$nativeVersion,\"$nativePath\");\n"
-        )
-
-        // read jni content.
-        val nativeDefineContent = readNativeMethods(file)
-        val methods = nativeDefineContent.split("},")
-
-        methods.forEach { methodLine ->
-            val list = methodLine.split(",")
-            if (list.size != 3) {
-                // println("error:${file.absolutePath}\n$nativeDefineContent")
-                return@forEach
-            }
-            val param1 = list[0]
-            val param2 = list[1]
-            val param3 = list[2] //  (void*)android_drm_DrmManagerClient_closeConvertSession
-
-            val jniMethod = param1.replace("\"", "").replace(" ", "")
-            //val jniMethodDescriptor = param2
-
-            val nativeMethod = param3.replace(" ", "").replace("(void*)", "")
-
-            var nativeMethodLine = -1
-            var jniLine = -1
-
-            file.readLines().forEachIndexed { index, text ->
-                if (text.contains(jniMethod)) {
-                    jniLine = index
-                }
-                val maybeRealDef1 = " $nativeMethod("
-                val maybeRealDef2 = " $nativeMethod ("
-                if ((text.contains(maybeRealDef1) || text.contains(maybeRealDef2)) && !text.contains(param3)) {
-                    nativeMethodLine = index
-                }
-                if (jniLine > 0 && nativeMethodLine > 0) {
-                    return@forEachIndexed
-                }
-            }
-
-            println(
-                "jni:$jniMethod descriptor:$param2 " +
-                    "jni-line:$jniLine nativeMethod:$nativeMethod nativeLine:$nativeMethodLine"
-            )
-
-            nativeMethodBuilder.append(
-                " INSERT INTO NativeMethodMapping " +
-                    "(file,jniMethod,nativeMethod,version,jniLine,defLine) VALUES (" +
-                    "\"$nativeKey\",\"$jniMethod\",\"$nativeMethod\",$nativeVersion,$jniLine,$nativeMethodLine" +
-                    ");\n"
-            )
-        }
-    }
-
-    private fun readNativeMethods(
-        file: File
-    ): String {
-        var beginIndex = Int.MAX_VALUE
-        var content = 0
-        val contentBuilder = StringBuilder()
-        file.readLines()
-            .filter { line -> line.isNotEmpty() }
-            .mapIndexed { index, line ->
-                line.trim().apply {
-                    if (contains("JNINativeMethod")) {
-                        beginIndex = index
-                        content = 1
-                    }
-                }
-            }
-            .forEachIndexed { index, line ->
-                // try to find method line
-                if (index > beginIndex) {
-                    // { "ctor", "(Ljava/io/FileDescriptor;)I", (void*)ctor_native },
-                    if (line.contains("{")) {
-                        content += 1
-                    }
-                    // read content.
-                    if (content > 1) {
-                        contentBuilder.append(line)
-                    }
-
-                    if (line.contains("}")) {
-                        content -= 1
-                        if (content < 1) {
-                            beginIndex = Int.MAX_VALUE
-                            return@forEachIndexed
-                        }
-                    }
-                }
-            }
-        return contentBuilder.toString()
-            .trim()
-            .replace("{", "")
-            .replace("\n", "")
     }
 }